# shared/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Wavelength API",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/connections/requests": {
      "get": {
        "summary": "Connection requests waiting on you.",
        "tags": [
          "connections"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "requests",
                    "count"
                  ],
                  "properties": {
                    "requests": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "user",
                          "message",
                          "createdAt",
                          "overallScore"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "user": {
                            "type": "object",
                            "required": [
                              "id",
                              "name",
                              "age",
                              "city",
                              "bio",
                              "photoUrl"
                            ],
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "age": {
                                "anyOf": [
                                  {
                                    "type": "integer"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "city": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "bio": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "photoUrl": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              }
                            }
                          },
                          "message": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "overallScore": {
                            "type": "integer"
                          }
                        }
                      }
                    },
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/connections/requests/count": {
      "get": {
        "tags": [
          "connections"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "count"
                  ],
                  "properties": {
                    "count": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/connections/requests/{id}/accept": {
      "post": {
        "tags": [
          "connections"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "matchId"
                  ],
                  "properties": {
                    "matchId": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/connections/requests/{id}/decline": {
      "post": {
        "tags": [
          "connections"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drops": {
      "get": {
        "summary": "The Drops feed.",
        "tags": [
          "drops"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "drops"
                  ],
                  "properties": {
                    "drops": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "required": [
                          "id",
                          "domain",
                          "itemKey",
                          "itemLabel",
                          "itemSubtitle",
                          "itemImage",
                          "caption",
                          "createdAt",
                          "author",
                          "community",
                          "likeCount",
                          "saveCount",
                          "likedByMe",
                          "savedByMe",
                          "sharedWithMe"
                        ],
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "domain": {
                            "type": "string"
                          },
                          "itemKey": {
                            "type": "string"
                          },
                          "itemLabel": {
                            "type": "string"
                          },
                          "itemSubtitle": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "itemImage": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "caption": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "author": {
                            "type": "object",
                            "required": [
                              "id",
                              "name",
                              "photoUrl"
                            ],
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "name": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              },
                              "photoUrl": {
                                "anyOf": [
                                  {
                                    "type": "string"
                                  },
                                  {
                                    "type": "null"
                                  }
                                ]
                              }
                            }
                          },
                          "community": {
                            "anyOf": [
                              {
                                "type": "object",
                                "required": [
                                  "id",
                                  "name",
                                  "slug"
                                ],
                                "properties": {
                                  "id": {
                                    "type": "string"
                                  },
                                  "name": {
                                    "type": "string"
                                  },
                                  "slug": {
                                    "type": "string"
                                  }
                                }
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "likeCount": {
                            "type": "integer"
                          },
                          "saveCount": {
                            "type": "integer"
                          },
                          "likedByMe": {
                            "type": "boolean"
                          },
                          "savedByMe": {
                            "type": "boolean"
                          },
                          "sharedWithMe": {
                            "type": "boolean"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Share something you like.",
        "tags": [
          "drops"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "domain",
                  "itemKey",
                  "itemLabel"
                ],
                "properties": {
                  "domain": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "music"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "movie"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "book"
                        ]
                      }
                    ]
                  },
                  "itemKey": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "itemLabel": {
                    "minLength": 1,
                    "type": "string"
                  },
                  "itemSubtitle": {
                    "type": "string"
                  },
                  "itemImage": {
                    "type": "string"
                  },
                  "caption": {
                    "maxLength": 400,
                    "type": "string"
                  },
                  "communityId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "id",
                    "domain",
                    "itemKey",
                    "itemLabel",
                    "itemSubtitle",
                    "itemImage",
                    "caption",
                    "createdAt",
                    "author",
                    "community",
                    "likeCount",
                    "saveCount",
                    "likedByMe",
                    "savedByMe",
                    "sharedWithMe"
                  ],
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "domain": {
                      "type": "string"
                    },
                    "itemKey": {
                      "type": "string"
                    },
                    "itemLabel": {
                      "type": "string"
                    },
                    "itemSubtitle": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "itemImage": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "caption": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "author": {
                      "type": "object",
                      "required": [
                        "id",
                        "name",
                        "photoUrl"
                      ],
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "name": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        },
                        "photoUrl": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      }
                    },
                    "community": {
                      "anyOf": [
                        {
                          "type": "object",
                          "required": [
                            "id",
                            "name",
                            "slug"
                          ],
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "name": {
                              "type": "string"
                            },
                            "slug": {
                              "type": "string"
                            }
                          }
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "likeCount": {
                      "type": "integer"
                    },
                    "saveCount": {
                      "type": "integer"
                    },
                    "likedByMe": {
                      "type": "boolean"
                    },
                    "savedByMe": {
                      "type": "boolean"
                    },
                    "sharedWithMe": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/drops/{id}/react": {
      "post": {
        "tags": [
          "drops"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "kind",
                  "on"
                ],
                "properties": {
                  "kind": {
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "like"
                        ]
                      },
                      {
                        "type": "string",
                        "enum": [
                          "save"
                        ]
                      }
                    ]
                  },
                  "on": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/drops/{id}": {
      "delete": {
        "tags": [
          "drops"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/communities/{id}/join": {
      "post": {
        "tags": [
          "communities"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}